Refactor: 재사용 가능성 낮은 hook 하나의 파일로 통합
