+++ src/components/inputs/rhf-rating-stars.tsx
@@ -1,3 +1,5 @@
+import { useMemo, useState } from 'react'
+import type { MouseEvent } from 'react'
 import { FieldValues, Path, useFormContext, useController } from 'react-hook-form'
 import { Star, StarHalf } from 'lucide-react'
 import {
@@ -10,7 +12,7 @@ import {
   ratingMessageActive,
   ratingStars,
 } from '@/styles/form-styles'
-import { useRatingLogic } from '@/features/book-form/hooks/use-rating-logic'
+import { clamp } from '@/utils/util'
 
 type Props<T extends FieldValues> = {
   name: Path<T>
@@ -21,7 +23,55 @@ type Props<T extends FieldValues> = {
   getDescription?: (value: number) => string | undefined
 }
 
+type UseRatingProps = {
+  value: unknown
+  onChange: (value: number) => void
+  onBlur: () => void
+  min: number
+  max: number
+  step: number
+}
+
 const STAR_COUNT = 5
+
+function useRating({ value, onChange, onBlur, min, max, step }: UseRatingProps) {
+  const [preview, setPreview] = useState<number | null>(null)
+  const segmentsPerStar = useMemo(() => Math.max(1, Math.round(1 / step)), [step])
+  const currentValue = typeof value === 'number' ? clamp(value, min, max) : null
+
+  const computeValue = (event: MouseEvent<HTMLButtonElement>, index: number) => {
+    const rect = event.currentTarget.getBoundingClientRect()
+    const ratio = clamp((event.clientX - rect.left) / rect.width, 0, 0.999)
+    const segment = Math.min(segmentsPerStar - 1, Math.floor(ratio * segmentsPerStar))
+    const base = min - step
+    const nextValue = base + index + step * segment
+    return clamp(parseFloat(nextValue.toFixed(2)), min, max)
+  }
+
+  const setDisplayValue = (event: MouseEvent<HTMLButtonElement>, index: number) => {
+    const next = computeValue(event, index)
+    onChange(next)
+    onBlur()
+    setPreview(null)
+  }
+
+  const setPreviewValue = (event: MouseEvent<HTMLButtonElement>, index: number) => {
+    setPreview(computeValue(event, index))
+  }
+
+  const setPreviewNull = () => {
+    setPreview(null)
+  }
+
+  const displayValue = preview ?? currentValue ?? 0
+
+  return {
+    displayValue,
+    setDisplayValue,
+    setPreviewValue,
+    setPreviewNull,
+  }
+}
 
 function iconType(value: number, starValue: number, step: number) {
   if (value >= starValue) return 'full'
@@ -43,7 +93,7 @@ export function RfhRatingStars<T extends FieldValues>({
     fieldState: { error },
   } = useController({ name, control })
 
-  const { displayValue, setDisplayValue, setPreviewValue, setPreviewNull } = useRatingLogic({
+  const { displayValue, setDisplayValue, setPreviewValue, setPreviewNull } = useRating({
     value: field.value,
     onChange: field.onChange,
     onBlur: field.onBlur,
