+++ src/utils/util.ts
@@ -1,0 +1,3 @@
+export function clamp(value: number, min: number, max: number) {
+  return Math.min(Math.max(value, min), max)
+}
